+++ phpmyfaq/admin/record.add.php
@@ -16,7 +16,6 @@
  * @since     2003-02-23
  */
 
-use Abraham\TwitterOAuth\TwitterOAuth;
 use phpMyFAQ\Administration\AdminLog;
 use phpMyFAQ\Category;
 use phpMyFAQ\Category\CategoryPermission;
@@ -32,7 +31,6 @@ use phpMyFAQ\Instance\Elasticsearch;
 use phpMyFAQ\Link;
 use phpMyFAQ\Notification;
 use phpMyFAQ\Question;
-use phpMyFAQ\Services\Twitter;
 use phpMyFAQ\Strings;
 use phpMyFAQ\Tags;
 use phpMyFAQ\Translation;
@@ -206,31 +204,6 @@ if ($user->perm->hasPermission($user->getUserId(), 'add_faq')) {
                         'category_id' => $categories['rubrik'][0]
                     ]
                 );
-            }
-
-            // Callback to Twitter if enabled
-            if ($faqConfig->get('socialnetworks.enableTwitterSupport')) {
-                $connection = new TwitterOAuth(
-                    $faqConfig->get('socialnetworks.twitterConsumerKey'),
-                    $faqConfig->get('socialnetworks.twitterConsumerSecret'),
-                    $faqConfig->get('socialnetworks.twitterAccessTokenKey'),
-                    $faqConfig->get('socialnetworks.twitterAccessTokenSecret')
-                );
-
-                $link = sprintf(
-                    'index.php?action=faq&amp;cat=%d&amp;id=%d&amp;artlang=%s',
-                    $categories['rubrik'][0],
-                    $recordId,
-                    $recordLang
-                );
-                $oLink = new Link($faqConfig->getDefaultUrl() . $link, $faqConfig);
-                $oLink->itemTitle = $question;
-                $link = $oLink->toString();
-
-                if ($connection) {
-                    $twitter = new Twitter($connection);
-                    $twitter->addPost($question, $tags, $link);
-                }
             }
 
             echo Alert::success('ad_entry_savedsuc');