+++ phpmyfaq/translations/language_en.php
@@ -997,14 +997,6 @@ $PMF_LANG['search.relevance.keywords-thema-content'] = 'Keywords - Question - An
 
 // added 2.7.0-alpha - 2010-09-13 by Thorsten
 $PMF_LANG['msgLoginUser'] = 'Login';
-$PMF_LANG['socialNetworksControlCenter'] = 'Social networks';
-$LANG_CONF['socialnetworks.enableTwitterSupport'] = ['checkbox', 'Twitter support'];
-$LANG_CONF['socialnetworks.twitterConsumerKey'] = ['input', 'Twitter Consumer Key'];
-$LANG_CONF['socialnetworks.twitterConsumerSecret'] = ['input', 'Twitter Consumer Secret'];
-
-// added 2.7.0-alpha - 2010-10-14 by Tom Zeithaml
-$LANG_CONF['socialnetworks.twitterAccessTokenKey'] = ['input', 'Twitter Access Token Key'];
-$LANG_CONF['socialnetworks.twitterAccessTokenSecret'] = ['input', 'Twitter Access Token Secret'];
 
 // added 2.7.0-alpha - 2010-12-21 by Anatoliy Belsky
 $PMF_LANG["ad_menu_attachments"] = "FAQ Attachments";
@@ -1151,7 +1143,6 @@ $PMF_LANG["ad_searchterm_del_suc"] = 'Successfully deleted all search terms.';
 $PMF_LANG["ad_searchterm_del_err"] = 'Could not delete all search terms.';
 $LANG_CONF["records.hideEmptyCategories"] = ["checkbox", "Hide empty categories"];
 $LANG_CONF["search.searchForSolutionId"] = ["checkbox", "Search for solution ID"];
-$LANG_CONF["socialnetworks.disableAll"] = ["checkbox", "Disable all social networks"];
 $LANG_CONF["main.enableGzipCompression"] = ["checkbox", "Enable GZIP compression"];
 
 // added v2.9.0-alpha2 - 2014-08-16 by Thorsten