+++ phpmyfaq/translations/language_es.php
@@ -997,14 +997,6 @@ $PMF_LANG['search.relevance.keywords-thema-content'] = 'Palabras clave - Pregunt
 
 // added 2.7.0-alpha - 2010-09-13 by Thorsten
 $PMF_LANG['msgLoginUser'] = 'Iniciar sesión';
-$PMF_LANG['socialNetworksControlCenter'] = 'Redes sociales';
-$LANG_CONF['socialnetworks.enableTwitterSupport'] = array('checkbox', 'Soporte de Twitter');
-$LANG_CONF['socialnetworks.twitterConsumerKey'] = array('input', 'Consumer Key de Twitter');
-$LANG_CONF['socialnetworks.twitterConsumerSecret'] = array('input', 'Consumer Secret de Twitter');
-
-// added 2.7.0-alpha - 2010-10-14 by Tom Zeithaml
-$LANG_CONF['socialnetworks.twitterAccessTokenKey'] = array('input', 'Access Token de Twitter');
-$LANG_CONF['socialnetworks.twitterAccessTokenSecret'] = array('input', 'Access TokenSecret de Twitter');
 
 // added 2.7.0-alpha - 2010-12-21 by Anatoliy Belsky
 $PMF_LANG['ad_menu_attachments'] = 'Adjuntos';