+++ phpmyfaq/translations/language_fr.php
@@ -997,14 +997,6 @@ $PMF_LANG["search.relevance.keywords-thema-content"] = "Mots-clés - Question - 
 
 // added 2.7.0-alpha - 2010-09-13 by Thorsten
 $PMF_LANG["msgLoginUser"] = "Connexion";
-$PMF_LANG["socialNetworksControlCenter"] = "Réseaux sociaux";
-$LANG_CONF["socialnetworks.enableTwitterSupport"] = ["checkbox", "Support Twitter"];
-$LANG_CONF["socialnetworks.twitterConsumerKey"] = ["input", "Twitter Consumer Key"];
-$LANG_CONF["socialnetworks.twitterConsumerSecret"] = ["input", "Twitter Consumer Secret"];
-
-// added 2.7.0-alpha - 2010-10-14 by Tom Zeithaml
-$LANG_CONF["socialnetworks.twitterAccessTokenKey"] = ["input", "Twitter Access Token Key"];
-$LANG_CONF["socialnetworks.twitterAccessTokenSecret"] = ["input", "Twitter Access Token Secret"];
 
 // added 2.7.0-alpha - 2010-12-21 by Anatoliy Belsky
 $PMF_LANG["ad_menu_attachments"] = "Pièces-jointes";