+++ phpmyfaq/translations/language_tr.php
@@ -988,14 +988,6 @@ $PMF_LANG['search.relevance.keywords-thema-content'] = "Anahtar sözcük - Soru 
 
 // added 2.7.0-alpha - 2010-09-13 by Thorsten
 $PMF_LANG['msgLoginUser'] = "Giriş";
-$PMF_LANG['socialNetworksControlCenter'] = "Sosyal ağlar";
-$LANG_CONF['socialnetworks.enableTwitterSupport'] = array('checkbox', 'Twitter support');
-$LANG_CONF['socialnetworks.twitterConsumerKey'] = array('input', 'Twitter Consumer Key');
-$LANG_CONF['socialnetworks.twitterConsumerSecret'] = array('input', 'Twitter Consumer Secret');
-
-// added 2.7.0-alpha - 2010-10-14 by Tom Zeithaml
-$LANG_CONF['socialnetworks.twitterAccessTokenKey'] = array('input', 'Twitter Access Token Key');
-$LANG_CONF['socialnetworks.twitterAccessTokenSecret'] = array('input', 'Twitter Access Token Secret');
 
 // added 2.7.0-alpha - 2010-12-21 by Anatoliy Belsky
 $PMF_LANG["ad_menu_attachments"] = "Eklentiler";
@@ -1128,7 +1120,6 @@ $PMF_LANG["ad_searchterm_del_suc"] = "Tüm arama terimleri başarıyla silindi."
 $PMF_LANG["ad_searchterm_del_err"] = "Arama terimleri silinemedi.";
 $LANG_CONF["records.hideEmptyCategories"] = array("checkbox", "Hide empty categories");
 $LANG_CONF["search.searchForSolutionId"] = array("checkbox", "Search for solution ID");
-$LANG_CONF["socialnetworks.disableAll"] = array("checkbox", "Disable all social networks");
 $LANG_CONF["main.enableGzipCompression"] = array("checkbox", "Enable GZIP compression");
 
 // added v2.9.0-alpha2 - 2014-08-16 by Thorsten